Rename and move object comparer and use it in isTracked and when updating entity SN-563
