+++ src/Saritasa.NetForge.Blazor/Controls/CustomFields/NavigationCollectionField.razor
@@ -1,4 +1,6 @@
-﻿@inherits CustomField
+﻿@using Saritasa.NetForge.DomainServices.Comparers
+
+@inherits CustomField
 
 @typeparam T
 
@@ -8,7 +10,7 @@
            HelperText="@Property.Description"
            Disabled="IsReadOnly"
            MultiSelection="true"
-           Comparer="new SelectComparer<T>()">
+           Comparer="new ObjectComparer<T>()">
     @foreach (var instance in NavigationInstances)
     {
         <MudSelectItem Value="instance" />
